ci: add GitHub Actions, and a deliberate ruff config
The repo had 254 passing tests that nothing ever ran automatically and no
linter at all -- the most visible gap for a project meant to be read.

CI runs backend and frontend as independent jobs so one stack's failure
does not mask the other's. Backend uses `uv sync --frozen`, which fails on
a lockfile that drifted from pyproject rather than silently resolving.
Frontend builds as well as lints, since with no TypeScript the build is
what actually catches a bad import.

The ruff config selects rule families rather than inheriting a default:
each one either caught something real here or guards a convention the code
already follows. Families that fight deliberate design are off, with the
reason recorded -- DTZ (market bars are naive dates by definition), S
(bandit heuristics on a local single-user app), RUF002/3 (the em-dash and
box-drawing comment style is house style, not homoglyph error).

Fixed rather than suppressed:
- B904: HTTPExceptions in route handlers now chain (`from e`), so a 500
  keeps the traceback that caused it. The two 401s use `from None`, since
  the caught NotAuthenticatedError adds nothing.
- SIM105, RUF046, B905/RUF007: contextlib.suppress, redundant int(round())
  casts, itertools.pairwise.

Suppressed with a stated reason, not blanketed:
- BLE001 at 8 sites, each annotated inline with why catching broadly is
  correct there. Route handlers get a per-file ignore, since converting any
  failure to a status code is that layer's entire job.
- B024 on MarketDataProvider: capabilities are declared as data, not by
  overriding methods, so abstract methods would defeat the design.

Verified locally exactly as CI runs it: uv sync --frozen, ruff clean,
259 passed, npm ci, eslint clean, vite build.

+++ backend/features/advisor/metrics.py
@@ -123,8 +123,8 @@ def trailing_return_pct(df: pd.DataFrame, months: float, skip_months: float = 0)
     `skip_months=1` gives the 12-1 momentum convention: the last month is
     dropped because short-term reversal contaminates it.
     """
-    lookback = int(round(months * TRADING_DAYS_PER_MONTH))
-    skip = int(round(skip_months * TRADING_DAYS_PER_MONTH))
+    lookback = round(months * TRADING_DAYS_PER_MONTH)
+    skip = round(skip_months * TRADING_DAYS_PER_MONTH)
     if lookback <= 0 or len(df) < lookback + skip + 1:
         return None
     closes = df["close"]
@@ -158,7 +158,7 @@ def reachability(
     (`atr% × √days`) and compares that budget to the target. Both inputs are the
     caller's — pass 5%/2mo and 10%/3mo and you get genuinely different answers.
     """
-    horizon_days = max(1, int(round(horizon_months * TRADING_DAYS_PER_MONTH)))
+    horizon_days = max(1, round(horizon_months * TRADING_DAYS_PER_MONTH))
     if daily_atr_pct is None or target_gain_pct <= 0:
         return {
             "horizon_trading_days": horizon_days,
